+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Changes in version 1.3.0 - 2018-??-??
+
+ * Medium changes
+   - Start downloading and processing votes.
+   - Add Apache Commons Math 3.6.1 as dependency.
+
+
 # Changes in version 1.2.0 - 2018-08-25
 
  * Medium changes